requirements.txt
streamlit
mysql-connector-python
langchain-core
langchain-groq
python-dotenv
